line-provider/requirements.txt
fastapi
uvicorn
aio-pika

pytest-asyncio
httpx